renamed some folders
