+++ src/domain/collaboration/calloutsSet/CalloutsInContext/CalloutsGroupView.tsx
@@ -1,7 +1,7 @@
 import CalloutCreationDialog from '../../callout/creationDialog/CalloutCreationDialog';
 import { useCalloutCreationWithPreviewImages } from '../useCalloutCreation/useCalloutCreationWithPreviewImages';
 import AddContentButton from '@/core/ui/content/AddContentButton';
-import CalloutsView, { CalloutsViewProps } from '../JourneyCalloutsTabView/CalloutsView';
+import CalloutsView, { CalloutsViewProps } from '../CalloutsView/CalloutsView';
 import { CalloutGroupName } from '@/core/apollo/generated/graphql-schema';
 import { useColumns } from '@/core/ui/grid/GridContext';
 import { useTranslation } from 'react-i18next';
